Also move unifyWithRename to prover package
Updated documentation for it, so it's much clearer what it does now
